Apply suggestions from code review

+++ src/theory/strings/proof_checker.cpp
@@ -141,7 +141,7 @@ Node StringProofRuleChecker::checkInternal(PfRule id,
       Node srem = utils::mkConcat(sremVec, stringType);
       return trem.eqNode(srem);
     }
-    // all remaining rules do something with the first side of each side
+    // all remaining rules do something with the first child of each side
     Node t0 = tvec[isRev ? nchildt - 1 : 0];
     Node s0 = svec[isRev ? nchilds - 1 : 0];
     if (id == PfRule::CONCAT_UNIFY)
@@ -151,7 +151,7 @@ Node StringProofRuleChecker::checkInternal(PfRule id,
       {
         return Node::null();
       }
-      for (unsigned i = 0; i < 2; i++)
+      for (size_t i = 0; i < 2; i++)
       {
         Node l = children[1][i];
         if (l.getKind() != STRING_LENGTH)
@@ -305,7 +305,7 @@ Node StringProofRuleChecker::checkInternal(PfRule id,
   {
     Assert(children.empty());
     Assert(args.size() >= 1);
-    // These rules are based on called a C++ method for returning a valid
+    // These rules are based on calling a C++ method for returning a valid
     // lemma involving a single argument term.
     // Must convert to skolem form.
     Node t = args[0];
@@ -463,7 +463,7 @@ Node StringProofRuleChecker::checkInternal(PfRule id,
       return Node::null();
     }
     Node t[2];
-    for (unsigned i = 0; i < 2; i++)
+    for (size_t i = 0; i < 2; i++)
     {
       if (children[0][i].getKind() == SEQ_UNIT)
       {
